refactor: complete pymech.viz subpackage migration
Complete the refactoring of visualization modules into the pymech.viz
subpackage with proper imports and updated documentation.

Changes:
- Fix relative imports in viz modules (..core, ..log instead of .core, .log)
- Update all test imports to use pymech.viz namespace
- Rename docs/pyvista_backend.rst to docs/viz.rst
- Update documentation to reference pymech.viz throughout
- Update API documentation references and examples
- Update docs/index.rst to reference viz.rst

The pymech.viz subpackage now properly exports:
- plot_mesh(), get_available_backends() - Main API
- hexa_to_pyvista(), add_boundary_conditions() - PyVista utilities
- MeshBackend, DEFAULT_BC_COLORS - Protocol and constants
- PyVistaBackend, MatplotlibBackend - Backend implementations

All tests pass and both backends (PyVista, Matplotlib) are working.
