+++ docs/index.rst
@@ -87,7 +87,7 @@ interpreter of your choice:
    neksuite.rst
    simsonsuite.rst
    vtksuite.rst
-   pyvista_backend.rst
+   viz.rst
    dataset.rst
    meshtools.rst
    usage.myst.md
